% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Streamliner.R
\name{catch_warnings_and_errors}
\alias{catch_warnings_and_errors}
\title{...}
\usage{
...
}
\arguments{
\item{expr}{The line(s) of R code that will be attempted}
}
\value{
...
}
\description{
...
}
\details{
...
The 'expr' parameter should be encased in curly braces ("{" and "}"), if it
consists of multiple R code expressions
}
\examples{
# ...
}
