% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Streamliner.R
\name{caught_issue}
\alias{caught_issue}
\title{...}
\usage{
...
}
\arguments{
\item{...}{...}
}
\value{
A single Boolean ('TRUE' if an error or warning occurred)
}
\description{
...
}
\details{
...
}
\examples{
# ...
}
