% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Key_Values.R
\name{cimis_start}
\alias{cimis_start}
\title{Get the Start Date for CIMIS}
\usage{
cimis_start()
}
\value{
A single \link{Date} value
}
\description{
This function returns a \link{Date} object containing the earliest date for which
daily data from \href{https://cimis.water.ca.gov/}{CIMIS} is available.
}
\details{
The \verb{California Irrigation Management Information System} (CIMIS) dataset
is an important component of SDA's hydrologic modeling work. Frequently in the
Russian River workflow, data before and after the start of CIMIS is used.
However, different conditions may apply based on the availability of CIMIS data.
Rather than type out the start date in every location, this function provides
easy access to that information.

In this package, the start date for CIMIS could have been stored as a variable
instead of a function, but because this package is stored within a sub-folder
of the larger repository, it is easier to track and update this value if it
is a function.
}
\examples{
# Define an example start date for a file
dataStart <- as.Date("2022-12-01", format = "\%Y-\%m-\%d") 

dataStart > cimis_start()
# Will be FALSE


# This function helps avoid typing out CIMIS's start date repeatedly
# across different functions and scripts
if (as.Date("1947-10-01", format = "\%Y-\%m-\%d") < cimis_start()) {
  #  This inner code would run
}
}
