% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model_Archive.R
\name{copy_file_to_archive}
\alias{copy_file_to_archive}
\title{Copy a File to a Model Archive Sub-Folder}
\usage{
copy_file_to_archive(inputPath, outputDirectory, 
                     model = "PRMS", subDir = "Input")
}
\arguments{
\item{inputPath}{A \link{character} vector containing one or more file paths.}

\item{outputDirectory}{A \link{character} string that contains the path to the
archive folder.}

\item{model}{A \link{character} string for the name of the model folder
(exactly as it appears in the archive folder).}

\item{subDir}{A \link{character} string that contains the name of the sub-folder
under \code{model} that will be the archive destination of the file.}
}
\value{
NULL (invisibly)
}
\description{
Given the path to a file, this function copies that file to an archive directory.
The same filename will be used, but the path will become
\code{outputDirectory/model/subDir/filename}.
}
\details{
\code{outputDirectory} contains the path to the archive folder. Within this folder,
there should be a folder with the same name as \code{model}. Inside the model folder,
there should be folders with names like "Input" and "Output". \code{subDir} should contain
the name of the planned destination folder under the model folder.

These three parameters help construct the planned destination path for the
input file. Using \code{\link[=extract_filename]{extract_filename()}}, the final piece is pulled from
\code{inputPath}. After that, \code{copyFile} (a function still contained within
the "Shared_Scripts" folder) is called to perform the file copy procedure.

Please note that the function is vectorized, so \code{inputPath} can contain one or
more filepaths. The destination folder will be the same in all cases.
}
\examples{
\dontrun{
# Example archive directory path
archiveDir <- "C:/Users/person/Archive"

# Example file
myFilePath <- "W2_Russian_River/Input/my_file.csv"


# Save the file to the "Output" folder under "PRMS" in 'archiveDir'
copy_file_to_archive(myFilePath, archiveDir, "PRMS", "Output")

# The copy's path will be "C:/Users/person/Archive/PRMS/Output/my_file.csv"


# Altneratively, save the file to the "Example" folder under "SRP" in 'archiveDir'
copy_file_to_archive(myFilePath, archiveDir, "SRP", "Example")

# The copy's path will be "C:/Users/person/Archive/SRP/Example/my_file.csv"
}
}
