% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Streamliner.R
\name{extract_filename}
\alias{extract_filename}
\title{Extract a File Name from a File Path}
\usage{
extract_file(path)
}
\arguments{
\item{path}{A \link{character} vector containing one or more file paths.}
}
\value{
A \link{character} \link{vector} with the same size as \code{path}.
}
\description{
This function is a shortcut for applying \code{\link[stringr:str_remove]{stringr::str_remove()}} to remove
all folders from a file path. For example, given a path like
"C:/Users/.../example.txt" or "C:\\Users\\...\\example.txt", this function
can return "example.txt".
}
\details{
This function relies on the default greedy behavior of pattern matching with
\code{\link[stringr:str_remove]{stringr::str_remove()}}. The regular expression in \code{extract_filename} matches
with everything in a string from its beginning to a backslash or forward slash.
With a \href{https://stackoverflow.com/questions/2301285/what-do-lazy-and-greedy-mean-in-the-context-of-regular-expressions}{greedy approach},
this will match everything in a file path until the final slash. The end result
of \code{\link[stringr:str_remove]{stringr::str_remove()}} will be the remnants of the string after that slash;
this should be just the filename.
}
\examples{
# Given a vector of paths, extract the filename portion
paths <- c("test.txt", "example/test.txt", "example_2\\\\test.txt")

# The function works with either a single string or multiple elements
extract_filename(paths[1])

extract_filename(paths)
}
