% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Streamliner.R
\name{find_matches}
\alias{find_matches}
\title{Find a Match in a Vector}
\usage{
find_matches(x, pattern, minMatches = 1, maxMatches = 1, 
             returnIndex = TRUE, filePath = NULL, 
             ignore.case = FALSE, fixed = FALSE)
}
\arguments{
\item{x}{A \code{\link[=character]{character()}} vector that will be searched for matches.}

\item{pattern}{A \code{\link[=character]{character()}} string that contains a regular expression. This
will be applied to \code{x} to search for matches.}

\item{minMatches}{An \code{\link[=integer]{integer()}} that identifies the minimum number of acceptable
matches that can be found by \code{pattern} in \code{x}.}

\item{maxMatches}{An \code{\link[=integer]{integer()}} that identifies the maximum number of acceptable
matches that can be found by \code{pattern} in \code{x}.}

\item{returnIndex}{A \code{\link[=logical]{logical()}} value that toggles whether a \code{\link[=numeric]{numeric()}} or
\code{\link[=character]{character()}} vector is returned by the function. If TRUE (the default), matching elements'
indices will be returned by the function. Otherwise, the function will return
the actual elements within \code{x} that match \code{pattern}. (This is the complement of
the \code{value} argument in \code{\link[=grep]{grep()}})}

\item{filePath}{An optional \code{character()} string (with default value \code{NULL}). This should
be the path to the file that is the source of \code{x}. See 'Details' for more information.}

\item{ignore.case}{A \code{\link[=logical]{logical()}} value that affects case sensitivity in the pattern
matching algorithm. Exactly like \code{\link[=grep]{grep()}}, this argument is FALSE by default,
causing \code{pattern} to be case sensitive in its search (e.g., "A" and "a" are
treated differently). If this argument is TRUE, case is ignored during matching
(e.g., "A" and "a" are equivalent).}

\item{fixed}{A \code{\link[=logical]{logical()}} value that determines whether \code{pattern} is treated like
a regular expression. Just like \code{\link[=grep]{grep()}}, this parameter is FALSE by default,
and \code{pattern} is applied in matching as a regular expression. Specify \code{TRUE} if
literal matches are desired (when TRUE, \code{fixed} can override other arguments).}
}
\value{
Either a \code{\link[=character]{character()}} or \code{\link[=integer]{integer()}} \code{\link[=vector]{vector()}}. Empty vectors are possible.
}
\description{
This function is a wrapper for the process of using \code{\link[=grep]{grep()}} to find matches
in a vector. It includes an optional error handling procedure related to the
number of matches found for a pattern.
}
\details{
\code{find_matches} helps find elements in a vector that satisfy a specified pattern
requirement. Just like \code{\link[=grep]{grep()}}, it can return either a numeric vector of indices
or a character vector with strings that match \code{pattern}.

Please note that, aside from \code{ignore.case}, \code{fixed}, and \code{value}, this function
relies on the default values of optional parameters in \code{\link[=grep]{grep()}}.

\code{minMatches} and \code{maxMatches} can help validate the results. With their default
values, a \code{pattern} should result in exactly one match in \code{x}. If not, the function
outputs an error message. To disable the error handling procedure, consider
setting \code{minMatches} and \code{maxMatches} to \code{-Inf} and \code{Inf}, respectively.

\code{filePath} is an optional argument that only affects error messages. If \code{x}
originated from a file, the path to that file can be specified here. Then, if
an error occurs, the output message will include this string. This can help
with debugging.
}
\examples{
# Given a vector of strings, find the index or indices that match a pattern
find_matches(month.abb, "o")

# To return the actual strings instead of their indices, use 'returnIndex'
find_matches(month.abb, "D", returnIndex = FALSE)

# Adjust expectations for results using 'minMatches' and 'maxMatches'
find_matches(month.abb, "e", minMatches = 3, maxMatches = 3)

# Disable error-checking entirely by setting 'minMatches' and 'maxMatches' to extreme values
find_matches(month.abb, "z", minMatches = -Inf, maxMatches = Inf)
}
