% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model_Archive.R
\name{get_archive_path}
\alias{get_archive_path}
\title{Get a Model Run's Associated Archive Folder}
\usage{
...
}
\arguments{
\item{...}{...}
}
\value{
...
}
\description{
Given the start date and end date bounds of the model run procedure as well
as the location of archive text files, this function locates this run's
specific file and retrieves the archive path from it.
}
\details{
Model procedure involves generating a text file
...
}
\examples{
\dontrun{
# Example archive directory path
...
}
}
