% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Control_Bridge.R
\name{get_from_master_control}
\alias{get_from_master_control}
\title{Get the Value for a Field in the Master Control File}
\usage{
get_from_master_control(fieldName)
}
\arguments{
\item{fieldName}{A \link{character} string exactly as it appears in the "FIELD"
column of the control file}
}
\value{
A single \link{character} string or similar value
}
\description{
To Be Written
}
\details{
To Be Written

Type of returned value is most likely a string
read_xlsx will assign the type for "VALUE" based on all of its values
If there's at least one text string in a row of that column, the entire
column will be read as \link{character}
}
\examples{
\dontrun{
# Extract the "VALUE" entry in the row that corresponds to the 
# "INITIAL_SHAREPOINT_FILE_PORTION" field
get_from_master_control("INITIAL_SHAREPOINT_FILE_PORTION")
}
}
