% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Key_Values.R
\name{prism_start}
\alias{prism_start}
\title{Get the Start Date for PRISM}
\usage{
prism_start()
}
\value{
A single \link{Date} value
}
\description{
This function returns a \link{Date} object containing the earliest date for which
daily data from \href{https://prism.oregonstate.edu/}{PRISM} is available.
}
\details{
The \verb{Parameter-elevation Regressions on Independent Slopes Model} (PRISM)
dataset is a core component of SDA's hydrologic modeling work. Frequently in the
Russian River workflow, data before and after the start of PRISM is used.
However, different conditions may apply based on the availability of PRISM data.
Rather than type out the start date in every location, this function provides
easy access to that information.

In this package, the start date for PRISM could have been stored as a variable
instead of a function, but because this package is stored within a sub-folder
of the larger repository, it is easier to track and update this value if it
is a function.
}
\examples{
# Define an example start date for a file
dataStart <- as.Date("2023-05-15", format = "\%Y-\%m-\%d") 

dataStart > prism_start()
# Will be FALSE


# This function helps avoid typing out PRISM's start date repeatedly
# across different functions and scripts
if (as.Date("1947-10-01", format = "\%Y-\%m-\%d") < prism_start()) {
  #  This inner code would run
}
}
