% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Control_Bridge.R
\name{read_master_control}
\alias{read_master_control}
\title{Read in the Master Control File as a Tibble}
\usage{
read_master_control()
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} containing the "Main" worksheet of the master control file
}
\description{
To Be Written
}
\details{
To Be Written

Wrapper for \code{getXLSX}, which in turn calls \code{\link[readxl:read_xlsx]{readxl::read_xlsx()}}
}
\examples{
\dontrun{
masterDF <- read_master_control()
}
}
