% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Web_Requests.R
\name{request_cdec}
\alias{request_cdec}
\title{Obtain data from CDEC}
\usage{
request_cdec(stationVec, startDate, endDate, sensorNum = 23, durCode = "D")
}
\arguments{
\item{stationVec}{A character vector containing one or more "Station IDs"
(exactly as they appear on CDEC's website).}

\item{startDate}{A \code{\link[=Date]{Date()}} object representing the first date in the data request.}

\item{endDate}{A \code{\link[=Date]{Date()}} object representing the last date in the data request.}

\item{sensorNum}{The type of variable to obtain data for. Each Sensor Number
corresponds to a different parameter. See a station's
\href{https://cdec.water.ca.gov/dynamicapp/staMeta?station_id=COY}{metadata page}
to determine which number should be specified. The default value is 23, which
corresponds to "RESERVOIR OUTFLOW, CFS".}

\item{durCode}{The timescale of the obtained data. "D" (the default) means
"Daily" data, "H" is for "Hourly", and "E" is for "Event".}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} containing the downloaded data.
}
\description{
This function can send a \code{\link[httr:GET]{httr::GET()}} request to the \href{https://cdec.water.ca.gov/index.html}{California Data Exchange Center} (CDEC) to procure data from
one or more stations.
}
\details{
Given a set of arguments, this function builds the required web request.
It has built-in error-handling capabilities and will process the result into
a \code{\link[tibble:tibble]{tibble::tibble()}}.

The columns of this result will match the data obtained from CDEC ("STATION_ID",
"DATE TIME", "VALUE", "UNITS", and "DATA_FLAG").

Typically, missing values are represented as "---" in the default data
returned by CDEC. However, these entries are detected and replaced with "-999".
This allows the column to be treated as numeric.
}
\examples{
\dontrun{
# Request data for Lake Sonoma ("WRS")
#
# According to its station metadata page, 
# it has "RESERVOIR INFLOW, CFS" data available on a daily scale 
# (https://cdec.water.ca.gov/dynamicapp/staMeta?station_id=WRS)
#
request_cdec("WRS", as.Date("2026-01-01"), Sys.Date(), sensorNum = 76, durCode = "D")


# Alternatively, if precipitation data from multiple stations is desired,
# this code can be executed instead
#
request_cdec(c("COY", "WRS"), as.Date("2023-10-01"), as.Date("2024-09-30"), 
             sensorNum = 45, durCode = "D")
}
}
